sincro

+++ src/main/java/com/jhernandez/backend/bazaar/application/service/CategoryService.java
@@ -6,10 +6,13 @@ import java.util.Optional;
 import com.jhernandez.backend.bazaar.application.port.CategoryRepositoryPort;
 import com.jhernandez.backend.bazaar.application.port.CategoryServicePort;
 import com.jhernandez.backend.bazaar.application.port.ImageServicePort;
+import com.jhernandez.backend.bazaar.application.port.ProductServicePort;
 import com.jhernandez.backend.bazaar.domain.exception.CategoryException;
 import com.jhernandez.backend.bazaar.domain.exception.ImageFileException;
+import com.jhernandez.backend.bazaar.domain.exception.ProductException;
 import com.jhernandez.backend.bazaar.domain.model.Category;
 import com.jhernandez.backend.bazaar.domain.model.ImageFile;
+import com.jhernandez.backend.bazaar.domain.model.Product;
 
 import lombok.RequiredArgsConstructor;
 
@@ -25,6 +28,7 @@ public class CategoryService implements CategoryServicePort{
 
     private final CategoryRepositoryPort categoryRepositoryPort;
     private final ImageServicePort imageServicePort;
+    private final ProductServicePort productServicePort;
 
     @Override
     public Optional<Category> createCategory(Category category, ImageFile categoryImage) throws CategoryException, ImageFileException {
@@ -79,10 +83,13 @@ public class CategoryService implements CategoryServicePort{
         }
 
     @Override
-    public void deleteCategoryById(Long id) throws CategoryException, ImageFileException {
+    public void deleteCategoryById(Long id) throws CategoryException, ImageFileException, ProductException {
         if (id == null) throw new CategoryException("Category ID must not be null");
+        if (id == 1L) throw new CategoryException("This category cannot be deleted");
         Category existingCategory = findCategoryById(id)
                 .orElseThrow(() -> new CategoryException("Category not found"));
+        List<Product> categoryProducts = productServicePort.findProductsByCategoryId(id);
+        productServicePort.removeProductCategoryList(categoryProducts, id);
         if (existingCategory.getImageUrl() != null) imageServicePort.deleteImageByUrl(existingCategory.getImageUrl());
         categoryRepositoryPort.deleteCategoryById(id);
     }
